requirements.txt
flask
bcrypt
pillow
markdown
meinheld
gunicorn
mysql-connector-python